Moved mpl HexTiles to its own module
