+++ holoviews/plotting/mpl/__init__.py
@@ -15,6 +15,7 @@ from .chart import * # noqa (API import)
 from .chart3d import * # noqa (API import)
 from .graphs import * # noqa (API import)
 from .heatmap import * # noqa (API import)
+from .hex_tiles import * # noqa (API import)
 from .path import * # noqa (API import)
 from .plot import * # noqa (API import)
 from .raster import * # noqa (API import)
